Mesh Outliner
Created a new game object and script to create an outline of the world mesh. Still some glitches, such as there still being edges inside the mesh having lines. Fortunately for the time being, these lines are hidden by the world mesh, as the lines are on a layer behind the world mesh.
The outliner could also use some performance improvements.