Find Points on Border of Mesh
Mesh outliner is now capable of finding all points on the border of a closed mesh.
Because of the way the slicer function operates, there are actually openings around the corners that are in the mesh's eyes "open".
To fix this, I will modify the slicing function to slice the mesh in such a way that there are no "openings" left.