(ONLY LAST SHADOW RENDERING) Multiple shadows masks from one compute shader.
All shadow masks are running on one shader, and therefore cutting down on loading the mesh information from the CPU to the GPU. This was a source of a ton of lag. Unfortunately, only the last Every DrawProceduralIndirect call only generates the shadow mask of the last light in the list of children. I am yet to figure out the cause of this, but that is the current problem I am working to solve.