SUPER SPEED per light shadow mask.
I fixed the lag problems by processing the mesh data to send to the compute shaders in the manager, rather than doing it once for every light. This cuts down on the load time drastically, and therefore many lights are now possible negligible (to me unnoticable lag).
Before I was trying to render every shadow mask with only one compute shader, but that wasn't working so I decided to backtrack to optimize this original solution. It's now working how I expected it too.