No Lag Spike
Lag spike was caused by the light manager translating a shared vertex mesh into an unshared vertex mesh. I figured out that this was unnecessary, and that a shared vertex mesh works just fine for the shadow mask compute shader.
The stress test revealed that lag spikes are no longer a problem for an arbitrarily large amount of shadow masks (256), but the consistent framerate decrease with the amount of lights. This is most likely due to the amount of triangles in the mesh generated. The most likely performance boost would be from only generating shadow masks from the triangles in a mesh that have edge-points.