Fixed Shadow Mask Offset from Main Camera.
The shadow mask no longer lags behind the main camera display. I thoughts this was due to lag in the post-processing lighting pipeline, but it was not.
The main camera was following the player using FixedUpdate, so the camera was moving between OnRenderImage calls.
This means that technically, there was no delay in the shadow mask. The real problem was that the main display was getting ahead of itself.